# (wip) feat: add padded locks for striped locking