--- a/libs/utils/padded_lock.go
+++ b/libs/utils/padded_lock.go
@@ -1,0 +1,15 @@
+package utils
+
+import (
+	"sync"
+
+	"golang.org/x/sys/cpu"
+)
+
+// PaddedLock is -as the name indicates- a lock which is padded in memory
+// exactly `cpu.CacheLinePad` length.
+// This is used for stripped locking to avoid-false sharing scenarios.
+type PaddedLock struct {
+	sync.Mutex
+	_ cpu.CacheLinePad
+}