## Overview

[Context](https://github.com/celestiaorg/celestia-node/pull/2612#discussion_r1322271835)

This PR introduces `PaddedLock` structured to avoid false sharing phenomena for striped locks

## TODO

- [ ] Include benchmarks

## Checklist

- [ ] New and updated code has appropriate documentation
- [ ] New and updated code has new and/or updated testing
- [ ] Required CI checks are passing
- [ ] Visual proof for any user facing features like CLI or documentation updates
- [ ] Linked issues closed with keywords