@derrandz 2023-09-29
@distractedm1nd Can you check up on the benchmark I added (_for starters_) for the eds store... the results I am finding don't show much improvement vs using regular mutexes, I find that weird. Would it be because it's eventually an I/O bound op and not a CPU op?

Will add benchmarks to the rest asap and see the difference, for now, here are the store's:

(_Tried a few homegrown parallelism approaches in benchmarking using wait groups and multiple routines before finally using `RunParallel`, not much difference_)

```
BenchmarkStore

Case 1)

- square size = 128
- parallelism = GOMAXPROCS

Padded stripped lock:
  BenchmarkStore/bench_put_128_parallel-2         	       2	 562190188 ns/op
  BenchmarkStore/bench_put_128_parallel-4         	       3	 472831820 ns/op
  BenchmarkStore/bench_put_128_parallel-8         	       2	 514809125 ns/op

Mutex stripped lock:

    BenchmarkStore/bench_put_128_parallel-2         	       2	 543366292 ns/op
    BenchmarkStore/bench_put_128_parallel-4         	       3	 470625778 ns/op
    BenchmarkStore/bench_put_128_parallel-8         	       3	 467839722 ns/op


- square size = 128
- parallelism = 200

PaddedLocks:

BenchmarkStore/bench_put_128_parallel-2         	       2	 541508938 ns/op
BenchmarkStore/bench_put_128_parallel-4         	       3	 485055959 ns/op
BenchmarkStore/bench_put_128_parallel-8         	       3	 468695945 ns/op

Mutex:

BenchmarkStore/bench_put_128_parallel-2         	       2	 535595584 ns/op
BenchmarkStore/bench_put_128_parallel-4         	       3	 462523208 ns/op
BenchmarkStore/bench_put_128_parallel-8         	       3	 475363972 ns/op


- square size = 128
- parallelism = 512

PaddedLocks:
BenchmarkStore/bench_put_128_parallel-2         	       2	 520852646 ns/op
BenchmarkStore/bench_put_128_parallel-4         	       3	 468282972 ns/op
BenchmarkStore/bench_put_128_parallel-8         	       2	 517932875 ns/op

Mutex:

BenchmarkStore/bench_put_128_parallel-2         	       2	 520831520 ns/op
BenchmarkStore/bench_put_128_parallel-4         	       3	 477697167 ns/op
BenchmarkStore/bench_put_128_parallel-8         	       3	 463459944 ns/op
```