@derrandz 2023-10-23
Update:

Here are benchmarks on an i9 machine. Although the theory behind this optimization makes sense, in practice, there isn't much performance gains. So I think I'll close this PR for now.

```
                                  │ benchmarks.mutex.txt │       benchmarks.padded.txt        │
                                  │        sec/op        │    sec/op     vs base              │
Store/bench_put_128-2                       1.856 ±  30%    1.898 ± 23%       ~ (p=0.818 n=6)
Store/bench_put_128-4                       1.788 ±  12%    1.761 ±  9%       ~ (p=0.818 n=6)
Store/bench_put_128-6                       2.021 ±  16%    1.834 ± 14%       ~ (p=0.310 n=6)
Store/bench_put_128-8                       2.148 ±  19%    2.036 ± 12%       ~ (p=0.485 n=6)
Store/bench_put_128-10                      2.194 ±  40%    2.179 ±  9%       ~ (p=0.699 n=6)
Store/bench_read_128-2                     290.8m ± 273%   334.4m ± 20%       ~ (p=0.818 n=6)
Store/bench_read_128-4                     197.9m ±  17%   187.4m ± 12%       ~ (p=0.240 n=6)
Store/bench_read_128-6                     147.9m ±   8%   151.9m ± 15%       ~ (p=0.699 n=6)
Store/bench_read_128-8                     120.4m ±  29%   123.2m ± 14%       ~ (p=0.589 n=6)
Store/bench_read_128-10                    127.3m ±   7%   130.5m ± 35%       ~ (p=0.240 n=6)
Store/bench_put_128_concurrent-2            2.103 ±  43%    2.511 ± 23%       ~ (p=0.180 n=6)
Store/bench_put_128_concurrent-4            1.334 ±  24%    1.345 ± 44%       ~ (p=0.818 n=6)
Store/bench_put_128_concurrent-6            1.113 ±  28%    1.115 ± 18%       ~ (p=0.818 n=6)
Store/bench_put_128_concurrent-8          1005.6m ±  23%   908.0m ± 20%       ~ (p=0.589 n=6)
Store/bench_put_128_concurrent-10          944.6m ±  13%   935.9m ±  4%       ~ (p=1.000 n=6)
geomean                                    745.6m          749.8m        +0.57%
```