@derrandz 2023-11-08
cc @distractedm1nd @Wondertan 

Will revaluate the actual need for this PR by providing an analysis using: https://grafana.com/docs/pyroscope/latest/configure-client/language-sdks/go_push/#mutex-profiling

This will be helpful in identifying all sources of contention within the app. This is going to be more insightful than just the theoretical reasons I've provided.

I will also revive this PR after getting some more accurate benchmarks on a server with a few hundered thousand runs... If pyroscope proves that there is a contention point indeed.